Read Me
I am a qualified and accomplished IT Quality Engineer with a BSc. in Information Technology from the Sri Lanka Institute of Information Technology (SLIIT), in addition to multiple professional qualifications and nearly 9 years’ combined work experience across the global SaaS industry—where I have served under various capacities, primarily in areas of Quality Assurance and Engineering, Project and Process Management, and Client Servicing for three of the leading global IT service providers based in Asia, Europe and Scandinavia, and the United States.